[savedObjects] change the number mappings to be 'long'
Since JavaScript numbers can't actually be integers we don't really care about
the difference between int and long, but since the long is the default type it
makes things more frictionless

+++ src/kibana/plugins/dashboard/services/_saved_dashboard.js
@@ -21,10 +21,10 @@ define(function (require) {
         // if type:dashboard has no mapping, we push this mapping into ES
         mapping: {
           title: 'string',
-          hits: 'integer',
+          hits: 'long',
           description: 'string',
           panelsJSON: 'string',
-          version: 'integer',
+          version: 'long',
           timeRestore: 'boolean',
           timeTo: 'string',
           timeFrom: 'string'
